Issue/556: feature: NPCs now driver system enabled
* Meaning, NPCs whose personality tab is toggled on now also display the ambition, aspirations and reactions input fields. So that bookkeeping of the agendas of important NPCs is now easier.
