+++ presentation/sheet/actor/part/personality/actor-personality.hbs
@@ -7,11 +7,13 @@ cssClass: {undefined | String}
   {{#with viewModel.personalityTraitsTemplate as | template |}}
     {{> (template) viewModel=../viewModel.personalityTraitsViewModel }}
   {{/with}}
-  {{#if viewModel.isPC}}
+  {{#if viewModel.displayDrivers}}
     {{#> header1 showFancyFont=viewModel.showFancyFont}}{{localize "system.character.driverSystem.label"}}{{/header1}}
     {{#with viewModel.driversTemplate as | template |}}
       {{> (template) viewModel=../viewModel.driversViewModel }}
     {{/with}}
+  {{/if}}
+  {{#if viewModel.isPC}}
     {{#> header1 showFancyFont=viewModel.showFancyFont}}{{localize "system.character.driverSystem.fateSystem.label"}}{{/header1}}
     {{#with viewModel.fateTemplate as | template |}}
       {{> (template) viewModel=../viewModel.fateViewModel }}
